backend/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
audioop-lts==0.2.2
audioread==3.1.0
Authlib==1.6.8
av==17.0.0
CacheControl==0.14.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
cryptography==46.0.4
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.3.1
fastapi==0.128.3
firebase_admin==7.2.0
frozenlist==1.8.0
google-adk==1.25.1
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-aiplatform==1.138.0
google-cloud-appengine-logging==1.8.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.40.1
google-cloud-bigquery-storage==2.36.2
google-cloud-bigtable==2.35.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-firestore==2.25.0
google-cloud-iam==2.21.0
google-cloud-logging==3.13.0
google-cloud-monitoring==2.29.1
google-cloud-pubsub==2.35.0
google-cloud-resource-manager==1.16.0
google-cloud-secret-manager==2.26.0
google-cloud-spanner==3.63.0
google-cloud-speech==2.36.1
google-cloud-storage==3.9.0
google-cloud-texttospeech==2.34.0
google-cloud-trace==1.18.0
google-crc32c==1.8.0
google-genai==1.62.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
graphviz==0.21
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.78.1
grpcio-status==1.78.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.26.0
mdurl==0.1.2
mmh3==5.2.0
msgpack==1.1.2
multidict==6.7.1
nest-asyncio==1.6.0
numba==0.64.0
numpy==2.4.2
openai==2.28.0
opentelemetry-api==1.39.1
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==26.0
platformdirs==4.9.2
pluggy==1.6.0
pooch==1.9.0
prometheus_client==0.24.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyOpenSSL==25.3.0
pyparsing==3.3.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.46
sqlalchemy-spanner==1.17.2
sqlparse==0.5.5
sse-starlette==3.2.0
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.52.1
tenacity==9.1.3
threadpoolctl==3.6.0
tqdm==4.67.3
typer==0.24.1
types-protobuf==6.32.1.20260221
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.23.0
zipp==3.23.0
